Add the ability to add derives on intermediates
In order to derive Insertable I need to add an attribute to handle
`table_name` and thus `intermediate_table_name` was born.

Hm... technically I ought to be able to use Diesel's own `table_name` and just
offer `intermediate_table_name` as an override.

This has not been verified with Associations, and it almost certainly doesn't work.